# go_web/docs/swagger.yaml
definitions:
  web.Response:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: string
    type: object
info:
  contact:
    name: API Support
    url: https://developers.mercadolibre.com.ar/support
  description: This API Handle MELI Products.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://developers.mercadolibre.com.ar/es_ar/terminos-y-condiciones
  title: MELI Bootcamp API
  version: "1.0"
paths:
  /transaction/:id:
    delete:
      consumes:
      - application/json
      description: delete transaction
      parameters:
      - description: id
        in: path
        name: tid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: transaction
      tags:
      - Transaction
    patch:
      consumes:
      - application/json
      description: patch transaction receptor
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Transaction
      tags:
      - Transaction
  /transactions/:id:
    put:
      consumes:
      - application/json
      description: put transaction
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Update Transaction
      tags:
      - Transaction
  /transactions/add:
    post:
      consumes:
      - application/json
      description: post transaction
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Add Transaction
      tags:
      - Transaction
  /transactions/get:
    get:
      consumes:
      - application/json
      description: get all transactions
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: List transactions
      tags:
      - Transaction
swagger: "2.0"